Rename DataArray header to data
