+++ xlive/Blam/Engine/Objects/Objects.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Blam\Math\BlamMath.h"
-#include "Blam\Engine\DataArray\DataArray.h"
+#include "Blam/Engine/memory/data.h"
 #include "Blam\Engine\Players\PlayerActions.h"
 #include "Blam\Engine\Objects\ObjectPlacementData.h"
