+++ xlive/H2MOD/Modules/MapManager/CustomMapDataCache/CustomMapDataCache.cpp
@@ -3,7 +3,7 @@
 #include "CustomMapDataCache.h"
 
 #include "H2MOD\Tags\TagInterface.h"
-#include "Blam\Engine\DataArray\DataArray.h"
+#include "Blam/Engine/memory/data.h"
 #include "Blam\FileSystem\FiloInterface.h"
 
 #include "Util\Hooks\Hook.h"
